% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-rst-first-arrival.R
\name{FirstArrivalTime}
\alias{FirstArrivalTime}
\alias{dft}
\alias{pft}
\alias{qft}
\alias{rft}
\title{The First Arrival Time Distribution}
\usage{
dft(x, alpha = 1, log = FALSE)

pft(q, alpha = 1, lower.tail = TRUE, log.p = FALSE)

qft(
  p,
  alpha = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  tol = .Machine$double.eps^0.25,
  max.steps = 1000
)

rft(n, alpha = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{alpha}{vector of (positive) parameters}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE, probabilities are P[X <= x], otherwise P[X > x]}

\item{p}{vector of probabilities}

\item{tol}{the tolerance for calculating quantiles}

\item{max.steps}{the maximum number of steps used to compute the quantiles}

\item{n}{number of random values to return}
}
\value{
dft gives the (log) density, pft gives the (log) distribution
  function, qft gives the quantile function, and rft generates random times.

  Invalid alpha will result in return value NaN, with a warning.

  The length of the result is determined by n for rft, and the length of the
  longest numeric argument for the other functions. The numerical arguments
  other than n are recycled to the length of the result. Only the first
  elements of the logical arguments are used.

  The returned values by qft are approximate, but are guaranteed to be
  within tol of the true value. A bisection search is used, so computation
  time increases with log(1/tol). If max.steps is reached, the current
  estimate is returned with a warning.
}
\description{
Density, distribution function, quantile function, and random number
generation for the first arrival time distribution with parameter alpha
}
\details{
The first arrival time distribution is defined as follows. If T ~ ft(alpha),
then T|S=s ~ uniform(0, s) where S ~ gamma(alpha + 1, alpha). It models the
distribution of the time until the first event in a gamma-count process after
a time L which is randomly chosen from the positive real numbers. (Let
L ~ uniform(0, M) and have M approach infinity.) The derivation and purpose
of this distribution can be found in the 'gammacount-derivation' vignette.
}
