% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-gamma-count.R
\name{GammaCount}
\alias{GammaCount}
\alias{dgc}
\alias{pgc}
\alias{qgc}
\alias{rgc}
\title{The Gamma-Count Distribution}
\usage{
dgc(x, lambda, alpha = 1, log = FALSE)

pgc(q, lambda, alpha = 1, lower.tail = TRUE, log.p = FALSE)

qgc(p, lambda, alpha = 1, lower.tail = TRUE, log.p = FALSE)

rgc(n, lambda, alpha = 1)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles}

\item{lambda}{vector of (positive) scale/mean parameters}

\item{alpha}{vector of (positive) dispersion parameters}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{q}{vector of quantiles}

\item{lower.tail}{logical; if TRUE, probabilities are P[X <= x], otherwise P[X > x]}

\item{p}{vector of probabilities}

\item{n}{number of random values to return}
}
\value{
dgc gives the (log) density, pgc gives the (log) distribution
  function, qgc gives the quantile function, and rgc generates random counts.

  Invalid lambda or alpha will result in return value NaN, with a warning.

  The length of the result is determined by n for rgc, and the length of the
  longest numeric argument for the other functions. The numerical arguments
  other than n are recycled to the length of the result. Only the first
  elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile function, and random number
generation for the gamma-count distribution with scale/mean parameter lambda
and dispersion parameter alpha
}
\details{
The gamma-count distribution models the count of events arriving in the
interval (0,lambda] when the times between successive events follow a
gamma(alpha, alpha) distribution. Since the expected time between events is
1, lambda is approximately the mean of the gamma-count distribution. This
distribution has the Poisson distribution as a special case when alpha = 1.
}
