% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-gamma-count-random-start-approx.R,
%   R/dist-gamma-count-random-start.R
\name{dgcrst.approx}
\alias{dgcrst.approx}
\alias{pgcrst.approx}
\alias{GammaCountRST}
\alias{dgcrst}
\alias{pgcrst}
\alias{qgcrst}
\alias{rgcrst}
\title{The Gamma-Count Distribution with Random Start Time}
\usage{
dgcrst.approx(x, lambda, alpha = 1, log = FALSE, segments = 10)

pgcrst.approx(
  x,
  lambda,
  alpha = 1,
  log.p = FALSE,
  lower.tail = TRUE,
  segments = 10
)

dgcrst(x, lambda, alpha = 1, log = FALSE, diagnostics = FALSE)

pgcrst(
  x,
  lambda,
  alpha = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  diagnostics = FALSE
)

qgcrst(p, lambda, alpha = 1, lower.tail = TRUE, log.p = FALSE)

rgcrst(n, lambda, alpha = 1)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles}

\item{lambda}{vector of (positive) scale/mean parameters}

\item{alpha}{vector of (positive) dispersion parameters}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{segments}{number of segments to use for approximate computation}

\item{lower.tail}{logical; if TRUE, probabilities are P[X <= x], otherwise P[X > x]}

\item{diagnostics}{logical; if TRUE, return diagnostics from integrate()}

\item{p}{vector of probabilities}

\item{n}{number of random values to return}

\item{q}{vector of quantiles}
}
\value{
dgcrst and dgcrst.approx give the (log) density, pgcrst and
  pgcrst.approx give the (log) distribution function, qgcrst gives the
  quantile function, and rgcrst generates random counts.

  Invalid lambda or alpha will result in return value NaN, with a warning.

  The length of the result is determined by n for rgcrst, and the length of the
  longest numeric argument for the other functions. The numerical arguments
  other than n are recycled to the length of the result. Only the first
  elements of the logical arguments are used.

  In dgcrst and pgcrst, if diagnostics=TRUE, the returned value is a list with
  components value, abs.error, subdivisions, message. All components are
  vectors with the same length. 'value' is the probability or density
  requested, and all other values come directly from any calls
  to integrate(). NA is put in the diagnostics for any value for which
  integrate() did not need to be called.
}
\description{
Density, distribution function, quantile function, and random number
generation for the gamma-count distribution with random start time with
scale/mean parameter lambda and dispersion parameter alpha
}
\details{
The gamma-count distribution with random start time models the count of
events arriving in the interval (L,L+lambda] where L is a large randomly
chosen positive real number and the times between successive events follow a
gamma(alpha, alpha) distribution. Since the expected time between events is
1, lambda is approximately the mean of the distribution. This
distribution has the Poisson distribution as a special case when alpha = 1.
The random start time improves the standard gamma-count distribution by
removing the behavior near t=0 for under- or over-dispersed values of alpha.

dgcrst.approx and pgcrst.approx use an approximation to the integral which
is relatively fast and still fairly accurate. The largest time savings is
with vector input.
}
\seealso{
[integrate()] for diagnostic output details, [dgc()] for the
  standard gamma-count distribution, and [dft()] for the first
  arrival time distribution after the random start time
}
