% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-gamma-count-random-start-joint.R
\name{GammaCountRSTJoint}
\alias{GammaCountRSTJoint}
\alias{dgcrst.joint}
\alias{rgcrst.joint}
\title{The Joint Gamma-Count RST and First Arrival Time Distribution}
\usage{
dgcrst.joint(x, tau, lambda, alpha = 1, log = FALSE)

rgcrst.joint(n, lambda, alpha = 1)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles}

\item{tau}{vector of (positive) first arrival time values}

\item{lambda}{vector of (positive) scale/mean parameters}

\item{alpha}{vector of (positive) dispersion parameters}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}

\item{n}{number of random values to return}
}
\value{
dgcrst.joint gives the (log) joint density, and rgcrst.joint
  generates random counts and arrival times.

  Invalid lambda or alpha will result in return value NaN, with a warning.

  The length of the result is determined by n for rgcrst.joint, and the
  length of the longest numeric argument for the other functions. The
  numerical arguments other than n are recycled to the length of the result.
  Only the first elements of the logical arguments are used.

  In dgcrst and pgcrst, if diagnostics=TRUE, the returned value is a list with
  components value, abs.error, subdivisions, message. All components are
  vectors with the same length. 'value' is the probability or density
  requested, and all other values come directly from any calls
  to integrate(). NA is put in the diagnostics for any value for which
  integrate() did not need to be called.
}
\description{
Density, and random number generation for the joint gamma-count distribution
with random start time and first arrival time, with scale/mean parameter
lambda and dispersion parameter alpha
}
\details{
The gamma-count distribution with random start time models the count of
events arriving in the interval (L,L+lambda] where L is a large randomly
chosen positive real number and the times between successive events follow a
gamma(alpha, alpha) distribution. The first arrival time is the time between
L and the first event which arrives after L, and may be greater than
L+lambda. The joint distribution of both (arising from the same process) is
given by these functions.
}
\seealso{
[dgc()] for the standard gamma-count distribution, and [dft()] for
  the first arrival time distribution after the random start time, and
  [dgcrst()] for the count distribution with random start time.
}
