% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-stationary-arrivals.R
\name{StationaryGammaArrival}
\alias{StationaryGammaArrival}
\alias{dsga}
\alias{psga}
\alias{qsga}
\alias{rsga}
\alias{dglel}
\alias{pglel}
\alias{qglel}
\alias{rglel}
\title{Stationary Gamma Arrival distribution}
\usage{
dsga(x, num = 1, alpha = 1, beta = alpha, log = FALSE)

psga(x, num = 1, alpha = 1, beta = alpha, lower.tail = TRUE, log.p = FALSE)

qsga(
  p,
  num = 1,
  alpha = 1,
  beta = alpha,
  lower.tail = TRUE,
  log.p = FALSE,
  tol = .Machine$double.eps^0.25,
  max.steps = 1000
)

rsga(n, num = 1, alpha = 1, beta = alpha)

dglel(x, alpha = 1, beta = alpha, log = FALSE)

pglel(x, alpha = 1, beta = alpha, lower.tail = TRUE, log.p = FALSE)

qglel(
  p,
  alpha = 1,
  beta = alpha,
  lower.tail = TRUE,
  log.p = FALSE,
  tol = .Machine$double.eps^0.25,
  max.steps = 1000
)

rglel(n, alpha = 1, beta = alpha)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{num}{vector of positive integers for the number of the arrival in
sequence. 1 is first, 2 is second, and so on.}

\item{alpha}{vector of (positive) shape parameters}

\item{beta}{vector of (positive) rate parameters. By default, beta = alpha
so that inter-arrival times have mean 1.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE, probabilities are P[X <= x], otherwise P[X > x]}

\item{p}{vector of probabilities}

\item{tol}{the tolerance for calculating quantiles}

\item{max.steps}{the maximum number of steps used to compute the quantiles}

\item{n}{number of random values to return}
}
\value{
dsga and dglel give the (log) density, psga and plel give the
  (log) distribution function, qsga and qlel give the quantile function, and
  rsga and rlel generate random times.

  Invalid alpha or beta will result in return value NaN.

  The length of the result is determined by n for rsga and rlel, and the
  length of the longest numeric argument for the other functions. The
  numerical arguments other than n are recycled to the length of the result.
  Only the first elements of the logical arguments are used.

  The returned values by qsga and qlel are approximate, but are guaranteed
  to be within tol of the true value. A bisection search is used, so
  computation time increases with -log(1/tol). If max.steps is reached, the
  current estimate is returned with a warning.
}
\description{
Density, distribution function, quantile function, and random number
generation for the arrival/renewal times in a stationary gamma renewal
process with parameters alpha and beta.
}
\details{
The stationary gamma renewal process is defined as follows: the first renewal
time is distributed according to the limiting excess life distribution in a
gamma renewal process. The times between all subsequent renewal times are
independently distributed gamma(alpha, beta).

These functions define the distribution of these arrival or renewal times,
the first of which is the limiting excess life distribution.

The 'glel' functions are shorthand for the 'sga' functions with num = 1.
}
